Replace Virtus with Vets::Model - MPI
